some cleaning introduce compare package
