+++ nccompare/compare/__init__.py
@@ -1,0 +1,1 @@
+from nccompare.compare.ncdiff import *
